+++ lib/componentt.dart
@@ -1,10 +1,17 @@
 library componentt;
 
+import 'dart:async';
+
 import 'package:flutter/widgets.dart';
 
 part 'src/fn.dart';
+
 part 'src/action.dart';
+
 part 'src/context.extension.dart';
+
 part 'src/component.dart';
+
 part 'src/control.dart';
+
 part 'src/scope.dart';